% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izzyuntappd.R
\docType{package}
\name{izzyuntappd}
\alias{izzyuntappd}
\alias{izzyuntappd-package}
\title{izzyuntappd: Datasets for Chester Ismay's Untappd beer ratings}
\description{
Package to load datasets from "Datasets for Mathematical Statistics
with Resampling in R" (2011) by Laura Chihara and Tim Hesterberg.
}
\section{Datasets}{

Type \code{data(package="izzyuntappd")} to view all datasets.
}

